# contracts/openapi/telemetry-api-v1.json
{"openapi":"3.1.0","info":{"title":"Industrial Telemetry Lab API","description":"Local ingestion, inventory, query, inspection, reconciliation, and replay API","license":{"name":"MIT","url":"https://opensource.org/license/mit"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Local Docker Compose gateway"}],"paths":{"/api/v1/admin/overload":{"put":{"tags":["admin-api"],"operationId":"overload","parameters":[{"name":"Authorization","in":"header","required":false,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OverloadRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"boolean"}}}}}}}},"/api/v1/admin/faults/outbox-confirm-gap":{"put":{"tags":["admin-api"],"operationId":"outboxConfirmGap","parameters":[{"name":"Authorization","in":"header","required":false,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OutboxConfirmGapRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"boolean"}}}}}}}},"/api/v1/replays":{"post":{"tags":["replay-api"],"operationId":"create","parameters":[{"name":"Authorization","in":"header","required":false,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReplayRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}}}}},"/api/v1/ingestion/batches":{"post":{"tags":["ingestion-api"],"operationId":"ingest","parameters":[{"name":"Authorization","in":"header","required":false,"schema":{"type":"string"}},{"name":"X-Lab-Fail-After-Raw-Store","in":"header","required":false,"schema":{"type":"boolean","default":false}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}}}}},"/api/v1/collectors/{collectorId}/heartbeat":{"post":{"tags":["collector-api"],"operationId":"heartbeat","parameters":[{"name":"collectorId","in":"path","required":true,"schema":{"type":"string"}},{"name":"Authorization","in":"header","required":false,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CollectorHeartbeat"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/v1/telemetry":{"get":{"tags":["query-api"],"operationId":"telemetry","parameters":[{"name":"facilityId","in":"query","required":true,"schema":{"type":"string"}},{"name":"assetId","in":"query","required":false,"schema":{"type":"string"}},{"name":"signalId","in":"query","required":false,"schema":{"type":"string"}},{"name":"from","in":"query","required":true,"schema":{"type":"string","format":"date-time"}},{"name":"to","in":"query","required":true,"schema":{"type":"string","format":"date-time"}},{"name":"includeFlagged","in":"query","required":false,"schema":{"type":"boolean","default":true}},{"name":"includeBadQuality","in":"query","required":false,"schema":{"type":"boolean","default":true}},{"name":"pageSize","in":"query","required":false,"schema":{"type":"integer","format":"int32"}},{"name":"cursor","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TelemetryPage"}}}}}}},"/api/v1/signals":{"get":{"tags":["query-api"],"operationId":"signals","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}}}}},"/api/v1/replays/{replayId}":{"get":{"tags":["replay-api"],"operationId":"get","parameters":[{"name":"replayId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}}}}},"/api/v1/collectors":{"get":{"tags":["collector-api"],"operationId":"all","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"type":"object","additionalProperties":{}}}}}}}}},"/api/v1/collectors/{collectorId}":{"get":{"tags":["collector-api"],"operationId":"one","parameters":[{"name":"collectorId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}}}}},"/api/v1/batches/{batchId}":{"get":{"tags":["query-api"],"operationId":"batch","parameters":[{"name":"batchId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}}}}},"/api/v1/admin/raw-objects/reconciliation":{"get":{"tags":["admin-api"],"operationId":"reconcile","parameters":[{"name":"Authorization","in":"header","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}}}}}},"components":{"schemas":{"OverloadRequest":{"type":"object","properties":{"overloaded":{"type":"boolean"}}},"OutboxConfirmGapRequest":{"type":"object","properties":{"armed":{"type":"boolean"},"batchId":{"type":"string","format":"uuid"}}},"ReplayRequest":{"type":"object","properties":{"facilityId":{"type":"string"},"from":{"type":"string","format":"date-time"},"to":{"type":"string","format":"date-time"},"mappingVersion":{"type":"string"},"qualityRulesVersion":{"type":"string"},"reason":{"type":"string"}}},"CollectorHeartbeat":{"type":"object","properties":{"collectorVersion":{"type":"string"},"configurationVersion":{"type":"string"},"sourceAdapterVersion":{"type":"string"},"spoolObservationCount":{"type":"integer","format":"int64"},"oldestUnsentObservationAgeSeconds":{"type":"integer","format":"int64"},"lastSuccessfulUploadAt":{"type":"string","format":"date-time"},"sourceConnected":{"type":"boolean"}}},"CanonicalSource":{"type":"object","properties":{"sourceSystem":{"type":"string"},"sourceEpoch":{"type":"string"},"sourceSequence":{"type":"integer","format":"int64"},"sourceTag":{"type":"string"},"collectorId":{"type":"string"},"collectorVersion":{"type":"string"},"mappingVersion":{"type":"string"},"qualityRulesVersion":{"type":"string"}}},"CanonicalTelemetrySample":{"type":"object","properties":{"contractVersion":{"type":"string"},"observationId":{"type":"string"},"facilityId":{"type":"string"},"assetId":{"type":"string"},"signalId":{"type":"string"},"observedAt":{"type":"string","format":"date-time"},"receivedAt":{"type":"string","format":"date-time"},"processedAt":{"type":"string","format":"date-time"},"value":{"type":"number","format":"double"},"unit":{"type":"string"},"quality":{"type":"string","enum":["GOOD","BAD"]},"flags":{"type":"array","items":{"type":"string"}},"source":{"$ref":"#/components/schemas/CanonicalSource"}}},"TelemetryPage":{"type":"object","properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/CanonicalTelemetrySample"}},"nextCursor":{"type":"string"}}}},"securitySchemes":{"localBearer":{"type":"http","description":"Synthetic local-development token used by mutating endpoints","scheme":"bearer"}}}}